% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup.R
\name{blup.FluMoDL}
\alias{blup.FluMoDL}
\alias{blup<-.FluMoDL}
\title{Get or set BLUP coefficients for a FluMoDL object}
\usage{
\method{blup}{FluMoDL}(m)

\method{blup}{FluMoDL}(m) <- value
}
\arguments{
\item{m}{An object of class \code{\link[=fitFluMoDL]{FluMoDL}}}
}
\value{
For \code{blup.FluMoDL}, the returned object of class
  \code{\link{summary.FluMoDL}} holding the BLUP coefficients associated
  with the FluMoDL object.
}
\description{
This retrieves or sets the BLUP coefficients for a particular
\link[=fitFluMoDL]{FluMoDL} object.
}
